add variables, improve text output

+++ import_database.sh
@@ -1,19 +1,32 @@
 #! /bin/bash
+V_DATABASE_USER_NAME=admin;
+V_DATABASE_USER_PASSWORD=secret;
+V_DIRECTORY_TO_SAVE=/home/mitaka/Desktop/;
 
-echo Enter database name:
+echo -n "Enter database name: "
 read i_database_name
 
-mysql -u USERNAME -pPASSWORD <<EOF
+echo -n "This operation will DROP database "\""${i_database_name}"\"" are you sure , say "\""yes"\"": " 
+read i_agreement 
+
+if [ "$i_agreement" != "yes" ]
+then
+	exit
+fi
+
+echo "Dropping database..."
+mysql -u ${V_DATABASE_USER_NAME} -p${V_DATABASE_USER_PASSWORD} <<EOF
 DROP DATABASE IF EXISTS ${i_database_name};
-
-CREATE DATABASE ${i_database_name}
- CHARACTER SET "utf8"
- COLLATE "utf8_general_ci"
 EOF
+echo "Dropped database."
 
-echo Created new database: "${i_database_name}"
-echo Begin importing...
-mysql -u USERNAME -pPASSWORD ${i_database_name} < /home/mitaka/Desktop/${i_database_name}.sql
-echo Done!
+mysql -u ${V_DATABASE_USER_NAME} -p${V_DATABASE_USER_PASSWORD} <<EOF
+CREATE DATABASE ${i_database_name}
+CHARACTER SET "utf8"
+COLLATE "utf8_general_ci"
+EOF
+echo "Created database."
 
-
+echo "Begin importing..."
+mysql -u ${V_DATABASE_USER_NAME} -p${V_DATABASE_USER_PASSWORD} ${i_database_name} < ${V_DIRECTORY_TO_SAVE}${i_database_name}.sql
+echo "Done."
